+++ bukkit/changelog.md
@@ -1,2 +1,3 @@
 - Don't allow inviting vanished players to groups
 - Improve Folia support
+- Added `threaded_server_support` config option